% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{convert_datetime_format}
\alias{convert_datetime_format}
\title{Convert EML date/time format string to one that R can parse}
\usage{
convert_datetime_format(eml_format_string)
}
\arguments{
\item{eml_format_string}{}
}
\value{
A date/time format string that can be parsed by \code{readr} or \code{strptime}.
}
\description{
Convert EML date/time format string to one that R can parse
}
\examples{
convert_datetime_format("MM/DD/YYYY")
}
