% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPchecker-package.R
\docType{package}
\name{DPchecker-package}
\alias{DPchecker}
\alias{DPchecker-package}
\title{DPchecker: Checks Data Packages for Congruence}
\description{
Allows the user (and reviewer) to check a data package and test whether it meets the congruence standards set forth by NPS for upload to DataStore as a datapackage.
}
\keyword{internal}
