% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{DPchecker_example}
\alias{DPchecker_example}
\title{Generate path to example data}
\usage{
DPchecker_example(dp_name = NULL)
}
\arguments{
\item{dp_name}{Name of data package. If omitted, this function will list all available example data packages.}
}
\value{
Path to example data, if dp_name is specified, or all available example data if not.
}
\description{
Generate path to example data
}
\examples{
DPchecker_example()
DPchecker_example("BUIS_herps")
}
