% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{is_eml}
\alias{is_eml}
\title{Check if metadata is eml object}
\usage{
is_eml(metadata)
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter not provided, defaults to calling \code{load_metadata} in current project directory.}
}
\description{
Helper function to validate that an argument belongs to the \code{emld} class.
}
