% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{load_data}
\alias{load_data}
\title{Load Data}
\usage{
load_data(directory = here::here())
}
\arguments{
\item{directory}{the directory where the data file(s) are found (i.e. your data package). Defaults to the current working directory. On exit, returns to the current working directory.}
}
\value{
a tibble of .csvs
}
\description{
load_data inspects the working directory for data files. Loads all existing data files into a tibble.
}
\details{
loads all data files in a specified directory (default is the working directory) into a tibble for later use in congruence checking. Returns the user to the working directory upon exit. Currently only supports .csv files.
}
\examples{
my_data <- load_data()
}
