% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{loadMetadata}
\alias{loadMetadata}
\title{Load Metadata}
\usage{
loadMetadata(directory = getwd())
}
\arguments{
\item{directory}{the directory where the metadata file is found (i.e. your data package). Defaults to the current working directory. On exit, returns you to the current working directory.}
}
\value{
an R-object formatted as EML metadata.
}
\description{
loads the metadata file from a given path or directory.
}
\details{

}
\examples{
my_metadata<-loadMetadata()

}
