% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{test.dupDataFiles}
\alias{test.dupDataFiles}
\title{Test Data Files for Duplicate Names}
\usage{
test.dupDataFiles(directory = getwd())
}
\arguments{
\item{directory}{path to your data files. Defaults to the current working directory.}
}
\description{
test.dupDataFiles looks at the file names of .csv files within a specified directory and tests whether there are any duplicates.
}
\details{
not sure why you'd ever need this function. Seems that most file systems won't let you have duplicate file names, but here it is for the sake of completeness. If you specify a directory other than your current working directory, it will return to the current working directory on exit.
}
\examples{
test.dupDataFiles()
}
