% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{test.fileNameMatch}
\alias{test.fileNameMatch}
\title{File Name Match}
\usage{
test.fileNameMatch(directory = getwd())
}
\arguments{
\item{directory}{path to the data files and metadata of a data package. Defaults to the current working directory.}
}
\value{
text
}
\description{
test.fileNameMatch checks to see whether all data files (.csv) within a specified directory are listed under the objectName (child of physical) element in an EML metadata file in the same directory, and vice versa. Mismatches will result in an error message.
}
\details{
If a directory other than the current working directory is specified, test.fileNameMatch returns to the current working directory on exit. Note that the metadata file must follow NPS naming conventions, specifically ending in *_metadata.xml. test.fileNameMatch assumes there are the same number of data files in the directory as dataTables in the metadata file.
}
\examples{
test.fileNameMatch()
}
