% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{test.footer}
\alias{test.footer}
\title{Footer Check}
\usage{
test.footer(emlObject)
}
\arguments{
\item{emlObject}{an R object containing EML metadata from a EML formatted .xml file loaded using either loadMetatdata() or EML::read_eml(<filename>, from="xml")}
}
\value{
message
}
\description{
test.footer checks the metadata files to determine whether data files contain footer lines or not.
}
\details{
If footer lines are not present, the data package passes the test. If footer lines are present, the data package fails the test and the user is instructed to remove footer lines prior to data package upload. Currently only EML metadata are supported.
}
\examples{
test.footer(my_metadata)
}
