% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_eml_elements.R
\name{test_attribute_defs}
\alias{test_attribute_defs}
\title{Test metadata for attribute definitions}
\usage{
test_attribute_defs(metadata = load_metadata(directory))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter not provided, defaults to calling \code{load_metadata} in current project directory.#'}
}
\value{
invisibly returns \code{metadata}
}
\description{
\code{test_attribute_defs()} extracts all attributeNames and attributeDefinitions from EML metadata. It tests to make sure there are the same number of attributeNames and attributeDefinitions. If true, the test passes. If not, it produces an error. Note bene: this test ONLY looks at the metadata, it does NOT look at the data files. Passing this test does not mean that all data columns have attributes in the metadata associated with them. To test that, see \code{test_fields_match()}.
}
\examples{
\dontrun{
test_attribute_defs()
}
}
