% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_eml_elements.R
\name{test_by_for_nps}
\alias{test_by_for_nps}
\title{Test for "by or for NPS"}
\usage{
test_by_for_nps(metadata = load_metadata(directory))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter is not provided, it defaults to calling \code{load_metadata} in the current project directory.#'}
}
\value{
invisibly returns \code{metadata}
}
\description{
Test for "by or for NPS"
}
\details{
\code{test_by_for_nps()} test for presence of the "by or for NPS" field. The test fails with an error if the information is missing from metadata. The test fails with a warning if the metadata indicate that the data were not created by or for the NPS (as this is expected to be relatively rare). Otherwise the test passes.
}
\examples{
\dontrun{
test_by_for_nps()
}
}
