% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_eml_elements.R
\name{test_content_units}
\alias{test_content_units}
\title{Test for content unit links}
\usage{
test_content_units(metadata = load_metadata(directory))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter is not provided, it defaults to calling \code{load_metadata} in the current project directory.}
}
\value{
invisible(metadata)
}
\description{
Test for content unit links
}
\examples{
\dontrun{
test_content_units()
}

}
