% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_eml_elements.R
\name{test_creator}
\alias{test_creator}
\title{Test for presence of creator and surName}
\usage{
test_creator(metadata = load_metadata(directory))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter is not provided, it defaults to calling \code{load_metadata} in the current project directory.}
}
\value{
invisible(metadata)
}
\description{
\code{test_creator()} examines the Creator element in metadata. If the creator element is missing, the test fails with an error. If the creator element is present, the function looks for individual creators. If individual creators are present and any of them lacks a surName, the test fails with an error. If any individual creators have a surName with more than two words, the test fails with a warning. Otherwise, the test passes.
}
\examples{
\dontrun{
test_creator()
}
}
