% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{test_datatable_url_attributes}
\alias{test_datatable_url_attributes}
\title{Test for the appropriate attribute for data table URLs in metadata}
\usage{
test_datatable_url_attributes(metadata = load_metadata(directory))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter not provided, defaults to calling \code{load_metadata} in current project directory.}
}
\value{
invisible(metadata)
}
\description{
\code{test_datatable_url_attributes} tests whether the 'function =' attribute for the \if{html}{\out{<url>}} element for each data table in metadata is properly specified. If there is no attribute, the function is assumed to be a direct download (as per the EML schema). The user is warned to check that this is the case (as data packages on DataStore will typically have a direct download link to the data file). If the specified attribute is not either "information" or "download", the function will throw an error as these are the only allowable attributes. If the attribute is "download" the function will warn the user and ask them to double check this. If the attribute is "information" and does not correspond to a DataStore reference profile, the function will warn the user and ask them to check this. If the attribute is "information" and a DataStore reference profile page is supplied, the test will pass.
}
\examples{
\dontrun{
test_datatable_url_attributes(metadata)
}
}
