% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{test_datatable_urls_doi}
\alias{test_datatable_urls_doi}
\title{Tests for data table URL formatting & correspondance with DOI}
\usage{
test_datatable_urls_doi(metadata = load_metadata(directory))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter not provided, defaults to calling \code{load_metadata} in current project directory.}
}
\value{
invisible(metadata)
}
\description{
\code{test_datatable_urls_doi()} passes if all data tables have URLs that are properly formatted (i.e. "https://irma.nps.gov/DataStore/Reference/Profile/xxxxxxx") where "xxxxxx" is identical to the DOI specified in the metadata. Fails with a warning if there is no DOI specified in metadata. If a DOI is specified in metadata, but the data table URL does not properly coincide with the url for the landing page that the doi points to for any one table, the test fails with a warning (and indicates which table failed). If data table urls do not exist, fails with an error and indicates how to add them.
}
\examples{
\dontrun{
dir <- DPchecker_example("BICY_veg")
test_datatable_urls_doi(dir)
}
}
