% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{test_delimiter}
\alias{test_delimiter}
\title{Field Delimiter Check}
\usage{
test_delimiter(directory = here::here())
}
\arguments{
\item{directory}{the directory where the metadata file is found - i.e. your data package. Defaults to the current working directory. On exit, returns you to the current working directory.}
}
\value{
message
}
\description{
test_delimiter checks the metadata file and ensures that each data file has a field delimiter with exactly one characte (e.g. ", ").
}
\details{
test_delimiter examines the fieldDelimiter element from EML (currently only EML is supported) metadata to determine how many characters there are. If there is no fieldDelimiter element, the test returns an error. If the field delimiter is anything other than exactly one character in length, the test returns an error.
}
\examples{
test_delimiter()
}
