% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{test_doi}
\alias{test_doi}
\title{Check for DOI}
\usage{
test_doi(metadata = load_metadata(directory))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter not provided, defaults to calling \code{load_metadata} in current project directory.}
}
\value{
Invisibly returns \code{metadata}.
}
\description{
\code{test_doi()} checks whether a DOI for the data package is present in metadata. It does not currently validate DOI. If a DOI is present, the test passes. If a DOI is not present, the test fails with a warning.
}
\examples{
meta <- load_metadata(DPchecker_example("BICY_veg"))
test_doi(meta)
}
