% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_eml_elements.R
\name{test_dp_abstract}
\alias{test_dp_abstract}
\title{Test EML abstract}
\usage{
test_dp_abstract(metadata = load_metadata(directory))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter is not provided, it defaults to calling \code{load_metadata} in the current project directory.}
}
\value{
invisibly returns \code{metadata}
}
\description{
\code{test_dp_abstract()} inspects EML for presence of a data package abstract. The test Fails with an error if the abstract is absent. If the abstract is present, the test fails with a warning if the abstract is <20 words, >250 words, or contains a subset of common characters that indicate improper formatting. Otherwise the test passes.
}
\examples{
\dontrun{
test_dp_abstract()
}
}
