% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{test_dup_data_files}
\alias{test_dup_data_files}
\title{Test Data Files for Duplicate Names}
\usage{
test_dup_data_files(directory = here::here())
}
\arguments{
\item{directory}{the directory where the data file(s) are found (i.e. your data package). Defaults to the current working directory. On exit, returns to the current working directory.}
}
\value{
message
}
\description{
test_dup_data_files looks at the file names of .csv files within a specified directory and tests whether there are any duplicates.
}
\details{
not sure why you'd ever need this function. Seems that most file systems won't let you have duplicate file names, but here it is for the sake of completeness. If you specify a directory other than your current working directory, it will return to the current working directory on exit.
}
\examples{
test_dup_data_files()

}
