% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{test_dup_meta_entries}
\alias{test_dup_meta_entries}
\title{Test Metadata for Duplicate Filenames}
\usage{
test_dup_meta_entries(metadata = load_metadata(here::here()))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter not provided, defaults to calling \code{load_metadata} in current project directory.}
}
\value{
Invisibly returns \code{metadata}.
}
\description{
test_dup_meta_entries test to see whether there are duplicate filenames listed for the data files in (EML) metadata.
}
\details{
specifically, test_dup_meta_entries looks at the 'physical' elements of a metadata file, which describe each data file, and asks whether there are duplicates entries under the objectName child element, which is where the file name for each data file is stored.
}
\examples{
meta <- load_metadata(DPchecker_example("BICY_veg"))
test_dup_meta_entries(meta)
}
