% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{test_field_num}
\alias{test_field_num}
\title{Test Number of Fields}
\usage{
test_field_num(directory = here::here(), metadata = load_metadata(directory))
}
\arguments{
\item{directory}{the directory where the data file(s) are found (i.e. your data package). Defaults to the current working directory. On exit, returns to the current working directory.}

\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter not provided, defaults to calling \code{load_metadata} in current project directory.}
}
\value{
message
}
\description{
test_field_num compares the number of attributes each dataTable within the EML metadata to the number of columns for the corresponding .csv. If the numbers are the same, the test passes. If the numbers differ, the test fails.
}
\details{
One thing to be cautious of: test.fieldNum does not compare the order or names of the columns! For now, that's on you.
}
\examples{
test_field_num()
}
