% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_eml_elements.R
\name{test_file_descript}
\alias{test_file_descript}
\title{Test presence of file descriptions}
\usage{
test_file_descript(metadata = load_metadata(directory))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter not provided, defaults to calling \code{load_metadata} in current project directory.#'}
}
\value{
invisibly returns \code{metadata}
}
\description{
\code{test_file_descriptions()} tests for the presence of file descriptions (\if{html}{\out{<entityDescription>}}) fields. It fails if any one or all of the \if{html}{\out{<entityDescription>}} fields are empty in a dataTable. It fails if any two or more file descriptions are identical. Otherwise the test passes. The test generates a warnign for each file description that is longer than 10 words or shorter than three words.
}
\examples{
\dontrun{
test_file_descript()
}
}
