% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_eml_elements.R
\name{test_file_descriptions}
\alias{test_file_descriptions}
\title{Test presence of file descriptions}
\usage{
test_file_descriptions(
  directory = here::here(),
  metadata = load_metadata(directory)
)
}
\arguments{
\item{directory}{the directory where the metadata file is found - i.e. your data package. Defaults to your current project directory.}

\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter not provided, defaults to calling \code{load_metadata} in current project directory.}
}
\value{
invisibly returns \code{metadata}
}
\description{
\code{test_file_descriptions()} tests for the presence of file descriptions (\if{html}{\out{<entityDescription>}}) fields. It tests that there are no duplications. It warns if descriptions are longer than 10 words and shorter than three words.
}
\examples{
\dontrun{
test_file_descriptions()
}
}
