% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{test_footer}
\alias{test_footer}
\title{Footer Check}
\usage{
test_footer(metadata = load_metadata(here::here()))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter not provided, defaults to calling \code{load_metadata} in current project directory.}
}
\value{
Invisibly returns \code{metadata}.
}
\description{
\code{test_footer()} checks the metadata files to determine whether data files contain footer lines or not.
}
\details{
If footer lines are not present, the data package passes the test. If footer lines are present, the data package fails the test with an error and the user is instructed to remove footer lines prior to data package upload. Currently only EML metadata are supported.
}
\examples{
meta <- load_metadata(DPchecker_example("BICY_veg"))
test_footer(meta)
}
