% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{test_geographic_cov}
\alias{test_geographic_cov}
\title{Check for Geographic Coverage
Checks if geographic coverage element is present in metadata. Does not perform any validation of geographic coverage information.}
\usage{
test_geographic_cov(metadata = load_metadata(directory))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter not provided, defaults to calling \code{load_metadata} in current project directory.}
}
\value{
Invisibly returns \code{metadata}.
}
\description{
Check for Geographic Coverage
Checks if geographic coverage element is present in metadata. Does not perform any validation of geographic coverage information.
}
\examples{
meta <- load_metadata(DPchecker_example("BICY_veg"))
test_geographic_cov(meta)
}
