% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{test_header_num}
\alias{test_header_num}
\title{Header Check}
\usage{
test_header_num(directory = here::here())
}
\arguments{
\item{directory}{the directory where the data file(s) are found (i.e. your data package). Defaults to the current working directory. On exit, returns to the current working directory.}
}
\value{
message
}
\description{
test_header_num checks the metadata files to ensure that each data file contains exactly one header row.
}
\details{
test_header_num examines the numHeaderLines element from EML (currently only EML is supported) metadata to determine how many header rows there are. If there are no header rows or if there is more than one header row, the test fails. The test also fails if there is no information about the number of header rows.
}
\examples{
test_header_num()
}
