% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_eml_elements.R
\name{test_license}
\alias{test_license}
\title{Test for presence of a license name}
\usage{
test_license(metadata = load_metadata(directory))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter not provided, defaults to calling \code{load_metadata} in current project directory.#'}
}
\value{
invisibly returns \code{metadata}
}
\description{
\code{test_license()} examines the licenseName element of EML metadata. If there is no license name, the test fails. If the license name does not match a list of valid license names, the test fails. If the metadata contain a valid license name, but the license name and CUI dissemination code do not agree, the test fails. Otherwise, the test passes.  Additionally, if the license name is not "Public Domain" or "CC0 1.0 Universal", the function throws a warning that the data package is not public.
}
\examples{
\dontrun{
test_license()
}
}
