% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{test_metadata_version}
\alias{test_metadata_version}
\title{EML Version Check}
\usage{
test_metadata_version(directory = getwd())
}
\arguments{
\item{directory}{the directory where the data file(s) are found (i.e. your data package). Defaults to the current working directory. On exit, returns to the current working directory.}
}
\value{
message
}
\description{
test_metadata_version determines whether the version of the metadata supplied meets the current criteria for an NPS data package.
}
\details{
currently only EML is supported. EML must be version >= 2.2.0.
}
\examples{
test_metadata_version()
}
