% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_eml_elements.R
\name{test_methods}
\alias{test_methods}
\title{Examines the Methods section of EML}
\usage{
test_methods(metadata = load_metadata(directory))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter is not provided, it defaults to calling \code{load_metadata} in the current project directory.}
}
\value{
invisible(metadata)
}
\description{
\code{test_methods()} first extracts the methods from an EML object. If methods are not present, the test fails with an error. If the methods are present, the tests asks 1) Is the section longer than 20 words? If not, the test fails with a warning.  2) Does the methods section contain unconventional characters such as &#13;? If so, the test fails with a warning. 2) Does the methods section contain additional spaces (more than two consecuitive spaces)? If so, the test fails with a warning. If all of the tests pass, the test as whole passes. For any error or warning, users are advised to use \code{EMLeditor::set_methods()} to correct the problem.
}
\examples{
\dontrun{
test_methods()
}
}
