% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_eml_elements.R
\name{test_notes}
\alias{test_notes}
\title{Examines the additionalInfo elment of EML metadata}
\usage{
test_notes(metadata = load_metadata(directory))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter is not provided, it defaults to calling \code{load_metadata} in the current project directory.}
}
\value{
invisible(metadata)
}
\description{
\code{test_notes()} extracts the additionalInfo components of EML metadata. These elements will be used to populate the "Notes" section on the DataStore landing page. If the Notes section is blank, the test fails with a warning. If the notes section contains non-standard characters (such as &#13;) or more than two consecuitive spaces, the test fails with a warning. Otherwise the test passes.  For all warnings, the user is advised to use \code{EMLeditor::set_additional_info()} to fix the additionalInfo section.
}
\examples{
\dontrun{
test_notes()
}
}
