% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{test_numeric_fields}
\alias{test_numeric_fields}
\title{Test Numeric Fields}
\usage{
test_numeric_fields(
  directory = here::here(),
  metadata = load_metadata(directory)
)
}
\arguments{
\item{directory}{the directory where the data file(s) are found (i.e. your data package). Defaults to the current working directory. On exit, returns to the current working directory.}

\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter not provided, defaults to calling \code{load_metadata} in current project directory.}
}
\value{
Invisibly returns \code{metadata}.
}
\description{
test_numeric_fields verifies that all columns listed as numeric in the metadata are free of non-numeric data.
}
\details{
"NA" and missing data codes documented in the metadata will \emph{not} cause this test to fail. Note that this test assumes that the column types that are in the metadata are the intended types, i.e., if your metadata says a column is text and it should actually be numeric, it will not be caught by this test.
}
\examples{
dir <- DPchecker_example("BICY_veg")
test_numeric_fields(dir)
}
