% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_eml_elements.R
\name{test_orcid_exists}
\alias{test_orcid_exists}
\title{Test creators for presence of an ORCiD}
\usage{
test_orcid_exists(metadata = load_metadata(directory))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter is not provided, it defaults to calling \code{load_metadata} in the current project directory.}
}
\value{
invisibly returns \code{metadata}
}
\description{
\code{test_orcid_exists()} will inspect the metadata and test each creator listed as an individual person (individualName) but not creators that are organizations for the presence of an ORCiD. If an ORCiD is found for all individual creators, the test passes. If any individual creator lacks an ORCiD, the test fails with a warning and the users is pointed towards \code{EMLeditor::set_creator_orcids()} to add ORCiDs if they so choose.
}
\examples{
\dontrun{
test_orcid_exists()
}
}
