% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_eml_elements.R
\name{test_orcid_match}
\alias{test_orcid_match}
\title{Tests whether metadata creator matches the ORCiD profile}
\usage{
test_orcid_match(metadata = load_metadata(directory))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter is not provided, it defaults to calling \code{load_metadata} in the current project directory.}
}
\value{
invisibly returns \code{metadata}
}
\description{
\code{test_orcid_match()} will only evaluate Creators that are individuals (not organizations). If an ORCiD has been supplied, the function will attempt to access the indicated ORCiD profile and test whether the last name indicated on the ORCiD profile matches the surName indicated in Metadata. If all surNames match the ORCiD profiles, the test passes. If any surName does not match the indicated ORCID profile, the test fails with an error.
}
\details{
Potential reasons for failing this test having entered the wrong ORCiD into metadata, having improperly formatted the ORCiD in metadata (it should be listed as xxxx-xxxx-xxxx-xxxx - see \code{test_orcid_format()}), having set your ORCiD profile to "private" (in which case the function can't access the name associated with the profile) or differences between the ORCiD profile name and the name in metadata (such as maiden vs. married name, transposing given and surnames, or variation in surName spelling).
}
\examples{
\dontrun{
test_orcid_match()
}
}
