% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_eml_elements.R
\name{test_orcid_resolves}
\alias{test_orcid_resolves}
\title{Test whether supplied Creator ORCiDs resolve to a valid ORCiD profile}
\usage{
test_orcid_resolves(metadata = load_metadata(directory))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter is not provided, it defaults to calling \code{load_metadata} in the current project directory.}
}
\value{
invisibly returns \code{metadata}
}
\description{
\code{test_orcid_resolves()} will only examine ORCiDs that are supplied for individual Creators (not organizations). If the ORCiD supplied consists of a URL that leads to a valid ORCiD profile, the test passes. If the ORCiD supplied is not a URL that resolves to a valid ORCiD profile - either because the ORCiD itself does not exist or because the ORCiD was supplied in an incorrect format, the test fails with an error. This test does not examine Creators that do not have associated ORCiDs; if no ORCiDs are provided that test does not return a pass or a fail.
}
\examples{
\dontrun{
test_orcid_resolves()
}
}
