% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_eml_elements.R
\name{test_project}
\alias{test_project}
\title{Test for a DataStore project}
\usage{
test_project(metadata = load_metadata(directory))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter is not provided, it defaults to calling \code{load_metadata} in the current project directory.}
}
\value{
invisible(metadata)
}
\description{
Test for a DataStore project
}
\examples{
\dontrun{
test_project()
}

}
