% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_eml_elements.R
\name{test_pub_date}
\alias{test_pub_date}
\title{Test publication date presence and ISO-8601 formatting}
\usage{
test_pub_date(metadata = load_metadata(directory))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter not provided, defaults to calling \code{load_metadata} in current project directory.#'}
}
\value{
invisibly returns \code{metadata}
}
\description{
tests for presence and ISO-8601 formatting of the publication date. Also tests whether the publication date is within a reasonable bounds; i.e. not a date prior to the beginning of the data package Reference Type (after 2022) and not in a future year.
}
\examples{
\dontrun{
meta <- load_metadata(DPchecker_example("BICY_Veg"))
test_pub_date(meta)
}
}
