% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{test_publisher}
\alias{test_publisher}
\title{Check for Publisher}
\usage{
test_publisher(metadata = load_metadata(directory), require_nps = FALSE)
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter not provided, defaults to calling \code{load_metadata} in current project directory.}

\item{require_nps}{If TRUE, throw an error if publisher information is not correct for NPS published data.}
}
\value{
Invisibly returns \code{metadata}.
}
\description{
\code{test_publisher()} checks if publisher information is present in metadata, with option to require valid NPS publisher information. If the publisher information is present, the test passes. If the publisher information is absent, the test fails with an error. If require_nps is set to TRUE (defaults to FALSE), the test will also ensure that a valid NPS publisher information is present. In this case, even if the publisher element is present, the test will fail with an error unless the publisher is the NPS (and all the publisher fields exactly match the expected information for NPS data packages).
}
\examples{
meta <- load_metadata(DPchecker_example("BICY_veg"))
test_publisher(meta)
}
