% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_eml_elements.R
\name{test_publisher_city}
\alias{test_publisher_city}
\title{Test EML metadata for publisher city - made redundant by test_publisher}
\usage{
test_publisher_city(
  directory = here::here(),
  metadata = load_metadata(directory)
)
}
\arguments{
\item{directory}{the directory where the metadata file is found - i.e. your data package. Defaults to your current project directory.}

\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter not provided, defaults to calling \code{load_metadata} in current project directory.}
}
\value{
invisibly returns \code{metadata}
}
\description{
Inspects the publisher address in EML metadata. Returns an error if the city element is empty. Returns a warning if it is not "Fort Collins". Passes test if it is "Fort Collins"
}
\examples{
\dontrun{
test_publisher_city()
}
}
