% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_eml_elements.R
\name{test_publisher_state}
\alias{test_publisher_state}
\title{Test EML metadata for publisher state - made redundant by test_publisher}
\usage{
test_publisher_state(metadata = load_metadata(directory))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter not provided, defaults to calling \code{load_metadata} in current project directory.#'}
}
\value{
invisibly returns \code{metadata}
}
\description{
Inspects the publisher address in EML metadata. Returns an error if the administrativeArea (state) element is empty. Returns a warning if it is not "CO". Passes test if it is "CO".
}
\examples{
\dontrun{
test_publisher_state()
}
}
