% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{test_valid_fieldnames}
\alias{test_valid_fieldnames}
\title{Test Field Names for Invalid Characters}
\usage{
test_valid_fieldnames(metadata = load_metadata(here::here()))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter not provided, defaults to calling \code{load_metadata} in current project directory.}
}
\value{
Invisibly returns \code{metadata}.
}
\description{
\code{test_valid_fieldnames()} checks for field names (e.g data column names) in the metadata that contain invalid special characters. Only underscores and alphanumeric characters are permitted, and names must begin with a letter. If invalid column names exist, the test will fail with a warning, otherwise the test passes.
}
\details{
You should run \code{test_fields_match()} before you run this function, since this function only checks the field names in the metadata.
}
\examples{
meta <- load_metadata(DPchecker_example("BICY_veg"))
test_valid_fieldnames(meta)
}
