% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{test_valid_filenames}
\alias{test_valid_filenames}
\title{Test File Names for Invalid Characters}
\usage{
test_valid_filenames(metadata = load_metadata(here::here()))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter not provided, defaults to calling \code{load_metadata} in current project directory.}
}
\value{
Invisibly returns \code{metadata}.
}
\description{
\code{test_valid_filenames()} checks for file names in the metadata that contain invalid special characters. Only underscores and alphanumeric characters are permitted, and names must begin with a letter. Currently, invalid filenames will result in the test failing with a warning, otherwise the test passes.
}
\details{
You should run \code{test_file_name_match()} before you run this function, since this function only checks the file names in the metadata.
}
\examples{
meta <- load_metadata(DPchecker_example("BICY_veg"))
test_valid_filenames(meta)
}
