% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{test_validate_schema}
\alias{test_validate_schema}
\title{Validate Metadata Schema}
\usage{
test_validate_schema(directory = here::here())
}
\arguments{
\item{directory}{the directory where the data file(s) are found (i.e. your data package). Defaults to the current working directory. On exit, returns to the current working directory.}
}
\value{
message
}
\description{
test_validate_schema inspects a metadata object loaded into R and determines whether it is schema-valid.
}
\details{
currently, only EML is supported. For now this is just a wrapper form EML::eml_validate().
}
\examples{
test_validate_schema()
}
