% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{validateSchema}
\alias{validateSchema}
\title{Validate Metadata Schema}
\usage{
validateSchema(emlObject)
}
\arguments{
\item{emlObject}{}
}
\description{
validateSchema inspects a metadata object loaded into R and determines whether it is schema-valid.
}
\details{
currently, only EML is supported. For now this is just a wrapper form EML::eml_validate().
}
