% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPchecker-package.R
\docType{package}
\name{DPchecker-package}
\alias{DPchecker}
\alias{DPchecker-package}
\title{DPchecker: Checks Data Packages for Congruence}
\description{
Allows the user (and reviewer) to check a data package and test whether it meets the congruence standards set forth by NPS for upload to DataStore as a datapackage.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nationalparkservice/DPchecker}
  \item \url{https://nationalparkservice.github.io/DPchecker/}
  \item Report bugs at \url{https://github.com/nationalparkservice/DPchecker/issues}
}

}
\author{
\strong{Maintainer}: Rob Baker \email{robert_baker@nps.gov} (\href{https://orcid.org/0000-0001-7591-5035}{ORCID})

Authors:
\itemize{
  \item Sarah E. Wright \email{sarah_wright@nps.gov}
}

Other contributors:
\itemize{
  \item Issac Quevedo [contributor]
  \item Amelia Sherman [contributor]
}

}
\keyword{internal}
