% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_checks.R
\name{run_congruence_checks}
\alias{run_congruence_checks}
\title{Run all congruence checks}
\usage{
run_congruence_checks(
  directory = here::here(),
  metadata = load_metadata(directory),
  check_metadata_only = FALSE,
  skip_cols = NA,
  output_filename,
  output_dir = here::here()
)
}
\arguments{
\item{directory}{the directory where the data file(s) are found (i.e. your data package). Defaults to the current working directory. On exit, returns to the current working directory.}

\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter not provided, defaults to calling \code{load_metadata} in current project directory.}

\item{check_metadata_only}{Only run checks on the metadata and skip anything involving data files.}

\item{skip_cols}{String. Defaults to NA. A list of one or more columns in the data to skip when testing whether the dates within data fall within the dates range specified in metadata. Useful if, for instance, there are columns within the data associated with the QA/QC process and these dates are expected to fall outside the date range specified for the data.}

\item{output_filename}{Optional. If specified, saves results of congruence checks to this file. If omitted, prints results to console. If the file already exists, results will be appended to the existing file.}

\item{output_dir}{Location in which to save the output file, if using.}
}
\value{
Invisibly returns \code{metadata}.
}
\description{
Run all congruence checks
}
\examples{
dir <- DPchecker_example("BICY_veg")
run_congruence_checks(dir)

}
