% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{run_congruence_checks}
\alias{run_congruence_checks}
\title{Run all congruence checks}
\usage{
run_congruence_checks(
  directory = here::here(),
  metadata = load_metadata(directory)
)
}
\arguments{
\item{directory}{the directory where the data file(s) are found (i.e. your data package). Defaults to the current working directory. On exit, returns to the current working directory.}

\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter not provided, defaults to calling \code{load_metadata} in current project directory.}
}
\value{
Invisibly returns \code{metadata}.
}
\description{
Run all congruence checks
}
\examples{
dir <- DPchecker_example("BICY_veg")
run_congruence_tests(dir)

}
