% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{test.headerNum}
\alias{test.headerNum}
\title{Header Check}
\usage{
test.headerNum(emlObject)
}
\arguments{
\item{emlObject}{an R object containing EML metadata from a EML formatted .xml file loaded using either loadMetatdata() or EML::read_eml(<filename>, from="xml")}
}
\description{
test.headerNum checks the metadata files to ensure that each data file contains exactly one header row.
}
\details{
headerCheck examines the numHeaderLines element from EML (currently only EML is supported) metadata to determine how many header rows there are. If there are no header rows or if there is more than one header row, the test fails. The test also fails if there is no information about the number of header rows.
}
\examples{
test.headerNum(my_metadata)
}
