% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{test.metadataVersion}
\alias{test.metadataVersion}
\title{EML Version Check}
\usage{
test.metadataVersion(emlObject)
}
\arguments{
\item{emlObject}{an R object containing EML metadata from a EML formatted .xml file loaded using either loadMetatdata() or EML::read_eml(<filename>, from="xml")}
}
\value{
message
}
\description{
test.metadataVersion determines whether the version of the metadata supplied meets the current criteria for an NPS data package.
}
\details{
currently only EML is supported. EML must be version >= 2.2.0.
}
\examples{
test.metadataVersion(my_metadata)
}
