% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{test.validateSchema}
\alias{test.validateSchema}
\title{Validate Metadata Schema}
\usage{
test.validateSchema(emlObject)
}
\arguments{
\item{emlObject}{an R object containing EML metadata from a EML formatted .xml file loaded using either loadMetatdata() or EML::read_eml(<filename>, from="xml")}
}
\value{
message
}
\description{
test.validateSchema inspects a metadata object loaded into R and determines whether it is schema-valid.
}
\details{
currently, only EML is supported. For now this is just a wrapper form EML::eml_validate().
}
\examples{
test.validateSchema(emlObject)
}
