% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_eml_elements.R
\name{test_cui_dissemination}
\alias{test_cui_dissemination}
\title{Test for CUI dissemination code}
\usage{
test_cui_dissemination(metadata = load_metadata(directory))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter is not provided, it defaults to calling \code{load_metadata} in the current project directory.}
}
\value{
invisibly returns \code{metadata}
}
\description{
\code{test_cui_dissemination()} examines EML metadata for the presence of a Controlled Unclassified Information (CUI) dissemination code. The function fails with an an error if the code does not exist or does not match a list of valid codes. If the valid code is not "PUBLIC" the test will produce a warning. A valid code results in a pass.
}
\examples{
\dontrun{
test_cui_dissemination()
}
}
