% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{test_date_range}
\alias{test_date_range}
\title{Test Date Range}
\usage{
test_date_range(directory = here::here(), metadata = load_metadata(directory))
}
\arguments{
\item{directory}{the directory where the data file(s) are found (i.e. your data package). Defaults to the current working directory. On exit, returns to the current working directory.}

\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter not provided, defaults to calling \code{load_metadata} in current project directory.}
}
\value{
Invisibly returns \code{metadata}.
}
\description{
test_date_range verifies that dates in the dataset are consistent with the date range in the metadata.
}
\details{
This function checks columns that are identified as date/time in the metadata. It throws a warning if the dates contained in the columns are outside of the temporal coverage specified in the metadata. If the date/time format string specified in the metadata does not match the actual format of the date in the CSV, it will likely fail to parse and throw an error.
}
\examples{
dir <- DPchecker_example("BICY_veg")
test_date_range(dir)
}
