% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{test_date_range}
\alias{test_date_range}
\title{Test Date Range}
\usage{
test_date_range(directory = here::here(), metadata = load_metadata(directory))
}
\arguments{
\item{directory}{the directory where the data file(s) are found (i.e. your data package). Defaults to the current working directory. On exit, returns to the current working directory.}

\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter not provided, defaults to calling \code{load_metadata} in current project directory.}
}
\value{
Invisibly returns \code{metadata}.
}
\description{
test_date_range verifies that dates in the dataset are consistent with the date range in the metadata.
}
\details{
Currently, this function operates on the assumption that the dates and times in your dataset are stored in a valid ISO format. If your dates/times are in a different format, this function may behave unpredictably.
}
\examples{
test_date_range()
}
