% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{test_doi}
\alias{test_doi}
\title{Check for DOI
Checks if DOI is present in metadata. Does not currently validate DOI.}
\usage{
test_doi(metadata = load_metadata(directory))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter not provided, defaults to calling \code{load_metadata} in current project directory.}
}
\value{
Invisibly returns \code{metadata}.
}
\description{
Check for DOI
Checks if DOI is present in metadata. Does not currently validate DOI.
}
\examples{
meta <- load_metadata(DPchecker_example("BICY_veg"))
test_geographic_cov(meta)
}
