% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_eml_elements.R
\name{test_dp_abstract}
\alias{test_dp_abstract}
\title{Test EML abstract}
\usage{
test_dp_abstract(metadata = load_metadata(directory))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter not provided, defaults to calling \code{load_metadata} in current project directory.#'}
}
\value{
invisibly returns \code{metadata}
}
\description{
inspects EML for presence of a data package abstract. FAils if abseent. If present, warns if the abstract is <20 words, >250 words, or contains a subset of common characters that indicate improper formatting.
}
\examples{
\dontrun{
test_dp_abstract()
}
}
