% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_eml_elements.R
\name{test_dp_title}
\alias{test_dp_title}
\title{Test data package title}
\usage{
test_dp_title(metadata = load_metadata(directory))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter not provided, defaults to calling \code{load_metadata} in current project directory.#'}
}
\value{
invisibly returns \code{metadata}
}
\description{
tests EML metadata for presence of a data package title. Warns if the title is > 15 words.
}
\examples{
\dontrun{
test_dp_title()
}
}
