% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_eml_elements.R
\name{test_dp_title}
\alias{test_dp_title}
\title{Test data package title}
\usage{
test_dp_title(metadata = load_metadata(directory))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter is not provided, it defaults to calling \code{load_metadata} in the current project directory.#'}
}
\value{
invisibly returns \code{metadata}
}
\description{
\code{test_dp_title()} tests EML metadata for presence of a data package title. The test fails with an error if the title is absent. The test fails with a warning if the title is > 15 words. Otherwise, the test passes.
}
\examples{
\dontrun{
test_dp_title()
}
}
