% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{test_field_num}
\alias{test_field_num}
\title{Test Number of Fields}
\usage{
test_field_num(directory = here::here())
}
\arguments{
\item{directory}{path to the data package files. Defaults to the current working directory.}
}
\value{
message
}
\description{
test_field_num compares the number of attributes each dataTable within the EML metadata to the number of columns for the corresponding .csv. If the numbers are the same, the test passes. If the numbers differ, the test fails.
}
\details{
One thing to be cautious of: test.fieldNum does not compare the order or names of the columns! For now, that's on you.
}
\examples{
test_field_num()
}
