% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_eml_elements.R
\name{test_file_descript}
\alias{test_file_descript}
\title{Test presence of file descriptions}
\usage{
test_file_descript(metadata = load_metadata(directory))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter is not provided, it defaults to calling \code{load_metadata} in the current project directory.#'}
}
\value{
invisibly returns \code{metadata}
}
\description{
\code{test_file_descript()} tests for the presence of file descriptions (entityDescription) fields. It fails with an error if any one or all of the entityDescription fields are empty in a dataTable. It fails with an error if any two or more file descriptions are identical. The test fails with a warning if for each file description that is longer than 15 words or shorter than three words. Otherwise the test passes.
}
\examples{
\dontrun{
test_file_descript()
}
}
