% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{test_footer}
\alias{test_footer}
\title{Footer Check}
\usage{
test_footer(directory = getwd())
}
\arguments{
\item{directory}{the directory where the data file(s) are found (i.e. your data package). Defaults to the current working directory. On exit, returns to the current working directory.}
}
\value{
message
}
\description{
test_footer checks the metadata files to determine whether data files contain footer lines or not.
}
\details{
If footer lines are not present, the data package passes the test. If footer lines are present, the data package fails the test and the user is instructed to remove footer lines prior to data package upload. Currently only EML metadata are supported.
}
\examples{
test_footer()
}
