% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_eml_elements.R
\name{test_int_rights}
\alias{test_int_rights}
\title{Test for presence of Intellectual Rights}
\usage{
test_int_rights(metadata = load_metadata(directory))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter not provided, defaults to calling \code{load_metadata} in current project directory.#'}
}
\value{
invisibly returns \code{metadata}
}
\description{
\code{test_int_rights()} test for the presence of text within the intellectualRights element in EML formatted metadata. If text if present, the test passes. Otherwise, the test fails. \code{test_int_rights()} makes no attempt to parse the text or test whether it properly coincides with the CUI dissemination codes or licenseName in the metadata. This is a simple presence/absence test.
}
\examples{
\dontrun{
test_int_rights()
}
}
