% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_eml_elements.R
\name{test_keywords}
\alias{test_keywords}
\title{Test for Keywords}
\usage{
test_keywords(metadata = load_metadata(directory))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter is not provided, it defaults to calling \code{load_metadata} in the current project directory.}
}
\value{
invisilbe(meatadatda)
}
\description{
\code{test_keywords()} tests to see whether metadata contains at least one "Keywords Set".
}
\examples{
\dontrun{
test_keywords()
}
}
