% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{test_metadata_version}
\alias{test_metadata_version}
\title{EML Version Check}
\usage{
test_metadata_version(metadata = load_metadata(here::here()))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter not provided, defaults to calling \code{load_metadata} in current project directory.}
}
\value{
Invisibly returns \code{metadata}.
}
\description{
\code{test_metadata_version()} determines whether the version of the metadata supplied meets the current criteria for an NPS data package.
}
\details{
currently only EML is supported. EML must be version >= 2.2.0.
}
\examples{
meta <- load_metadata(DPchecker_example("BICY_veg"))
test_metadata_version(meta)
}
