% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{test_publisher}
\alias{test_publisher}
\title{Check for Publisher
Checks if publisher information is present in metadata, with option to require valid NPS publisher information.}
\usage{
test_publisher(metadata = load_metadata(directory), require_nps = FALSE)
}
\arguments{
\item{require_nps}{If TRUE, throw an error if publisher information is not correct for NPS published data.}
}
\value{
Invisibly returns \code{metadata}.
}
\description{
Check for Publisher
Checks if publisher information is present in metadata, with option to require valid NPS publisher information.
}
\examples{
meta <- load_metadata(DPchecker_example("BICY_veg"))
test_publisher(meta)
}
