% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_eml_elements.R
\name{test_publisher_city}
\alias{test_publisher_city}
\title{Test EML metadata for publisher city}
\usage{
test_publisher_city(metadata = load_metadata(directory))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter is not provided, it defaults to calling \code{load_metadata} in the current project directory.#'}
}
\value{
invisibly returns \code{metadata}
}
\description{
\code{test_publisher_city()} inspects the publisher address in EML metadata. It fails with an an error if the city element is empty. It fails with a warning if it is not "Fort Collins". The test passes if it is "Fort Collins"
}
\examples{
\dontrun{
test_publisher_city()
}
}
