% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_eml_elements.R
\name{test_publisher_name}
\alias{test_publisher_name}
\title{Tests EML metadata for the publisher name}
\usage{
test_publisher_name(
  directory = here::here(),
  metadata = load_metadata(directory)
)
}
\arguments{
\item{directory}{the directory where the metadata file is found - i.e. your data package. Defaults to your current project directory.}

\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter not provided, defaults to calling \code{load_metadata} in current project directory.}
}
\value{
invisibly returns \code{metadata}
}
\description{
Tests EML metadata for the publisher name
}
\details{
error if the publisher name is missing, warning if the publisher name is not "National Park Service". Passes test if publisher name is "National Park Service"
}
\examples{
\dontrun{
test_publishername()
}
}
