% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_eml_elements.R
\name{test_publisher_name}
\alias{test_publisher_name}
\title{Tests EML metadata for the publisher name}
\usage{
test_publisher_name(metadata = load_metadata(directory))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter is not provided, it defaults to calling \code{load_metadata} in the current project directory.}
}
\value{
invisibly returns \code{metadata}
}
\description{
Tests EML metadata for the publisher name
}
\details{
\code{test_publisher_name()} test for the presence of the data package publisher name. It fails with an error if the publisher name is missing, and fails with a warning if the publisher name is not "National Park Service" (as this is expected to be rare). Passes test if publisher name is "National Park Service"
}
\examples{
\dontrun{
test_publisher_name()
}
}
