% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{test_taxonomic_cov}
\alias{test_taxonomic_cov}
\title{Check for Taxonomic Coverage}
\usage{
test_taxonomic_cov(metadata = load_metadata(directory))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter not provided, defaults to calling \code{load_metadata} in current project directory.}
}
\value{
Invisibly returns \code{metadata}.
}
\description{
'test_taxnomomic_cov()` checks whether taxonomic coverage element is present in metadata. It does not perform any validation of taxonomic coverage information. If taxonomic coverage is present, the test passes. If it is absent, the test fails with a warning.
}
\examples{
meta <- load_metadata(DPchecker_example("BICY_veg"))
test_taxonomic_cov(meta)
}
