% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_data_congruence.R
\name{test_validate_schema}
\alias{test_validate_schema}
\title{Validate Metadata Schema}
\usage{
test_validate_schema(metadata = load_metadata(here::here()))
}
\arguments{
\item{metadata}{The metadata object returned by \code{load_metadata}. If parameter not provided, defaults to calling \code{load_metadata} in current project directory.}
}
\value{
Invisibly returns \code{metadata}.
}
\description{
test_validate_schema inspects a metadata object loaded into R and determines whether it is schema-valid.
}
\details{
currently, only EML is supported. For now this is just a wrapper form EML::eml_validate().
}
\examples{
meta <- load_metadata(DPchecker_example("BICY_veg"))
test_validate_schema(meta)
}
